Giorgionocera/0.0.1 docs adding docs for modules (#143)
* docs: :memo: restore fantoken module doc

Restoring the fantoken module documentation from an old repository.

* docs(x): :memo: add link to merkledrop docs

Adding a link to the merkledrop module documentation (readme) on the "modules documentation".

* docs(x/fantoken): :memo: update readme for fantoken module

Updating the README doc for fantoken module documentation for adapting to the new implementation.

* docs: :memo: update concepts to match new implementation

Updating the concepts document to match new implementation of the Fantoken module.

* docs: :memo: update state doc for fantoken

Updating the state document for the fantoken module documentation to match the new implementation.

* docs: :memo: fix docs missing information on uri

Fixing the concepts document about the fantoken module documentation which was missing about information on uri length.

* docs(x/fantoken): :memo: update lifecycle of a fantoken

Updating the lifecycle of a fantoken on the fantoken module documentation.

* docs(x/fantoken): :memo: fix docs

Fixing documentation for fantokens concepts.

* docs(x/fantoken): :memo: fix spacing in documentation

Fixing spacing in the fantoken module documentation.

* docs(x/fantoken): :memo: update doc to meet new implementation

Updating the messages document to meet new implementation in the fantoken module documentation.

* docs(x/fantoken): :memo: update events in fantoken module

Updating the events document in the fantoken module documentation to meet the new implementation.

* docs: :memo: update parameters in the fantoken module docs

Updating the parameters document in the fantoken module documentation to meet the new implementation.

* docs(x/fantoken): :memo: update client documentation for fantoken module

Updating the client document for fantoken module documentation to meet the new implementation.

* docs(x/fantoken): :memo: update fantoken module readme

Updating the fantoken module documentation.

* docs(x/fantoken): :memo: update doc fantoken module

Updating the fantoken module documentation.

* docs(x/fantoken): :memo: update fantoken module docs

Updating the fantoken module documentation.

* docs(x/fantoken): :memo: update ToC in doc

Updating the Table of Content in the fantoken module documentation.

* docs(x/fantoken): :memo: fix typo in doc

Fixing typos in fantoken module documentation.

* docs(x/merkledrop): :memo: add events, parameters, client docs for merkledrop module

Adding events, parameters, client documents for merkledrop module documentation.

* docs(x/merkledrop): :memo: wip messages doc merkledrop

Work in progress for the messages document in the merkledrop module documentation.

* docs(x/merkledrop): :memo: WIP messages doc in merkledrop module

Work in progress add content to messages document in the merkledrop module documentation.

* docs(x/merkledrop): :memo: WIP messages completed MsgCreate

Work in progress in the messages document. Completed the MsgCreate documentation in the messages.

* docs(x/merkledrop): :memo: add messages document to merkledrop module doc

Adding the messages document to merkledrop module documentation.

* docs(x/merkledrop): :memo: add end_block doc in merkledrop module

Adding the end_block document to the merkledrop module documentation.

* docs(x/merkledrop): :memo: add images for docs

Adding images for merkledrop module documentation.

* docs(x/merkledrop): :memo: fix images for doc

Fixing images for merkledrop module documentation.

* docs(x/merkledrop): :memo: fix typo in docs

Fixing typo in the merkledrop module documentation.

* docs(x/merkledrop): :memo: add concepts for merkledrop documentation

Adding the concepts document for the merkledrop module documentation.

* docs(x/merkledrop): :memo: fix typo in docs

Fixing typos in the merkledrop docs.

* docs(x/merkledrop): :memo: add state and readme for merkledrop module

Adding a state document and a README for merkledrop module documentation.

* docs(x/merkledrop): :memo: fix toc in merkledrop module

Fixing the Table of Content in the merkledrop module documentation.